+++ src/notebook/ipython_profiles/profile_narrative/kbase_templates/static/kbase/js/kbaseNarrative.js
@@ -121,32 +121,39 @@ EndpointTester.prototype.test = function(url) {
  */
 var _kb_failed_once = false;
 var KBFail = function(is_fatal, where, what) {
-  if (!IPython || !IPython.notebook) {
-    return false;
-  }
-  var code = "";
-  if (_kb_failed_once == false) {
-    code += "from biokbase.narrative.common import kblogging\n";
-    _kb_failed_once = true;
-  }
-  code += "kblogging.NarrativeUIError(";
-  if (is_fatal) {
-    code += "True,";
-  }
-  else {
-    code += "False,";
-  }
-  if (where) {
-    code += 'where="' + where + '"';
-  }
-  if (what) {
-    if (where) { code += ", "; }
-    code += 'what="' + what + '"';
-  }
-  code += ")\n";
-  // Log the failure
-  IPython.notebook.kernel.execute(code);
-  return true;
+    if (!IPython || !IPython.notebook || !IPython.notebook.kernel) {
+        return false;
+    }
+    var code = "";
+    if (_kb_failed_once == false) {
+        code += "from biokbase.narrative.common import kblogging\n";
+        _kb_failed_once = true;
+    }
+    code += "kblogging.NarrativeUIError(";
+    if (is_fatal) {
+        code += "True,";
+    }
+    else {
+        code += "False,";
+    }
+    if (where) {
+        code += 'where="' + where + '"';
+    }
+    if (what) {
+        if (where) { code += ", "; }
+        code += 'what="' + what + '"';
+    }
+    code += ")\n";
+    // Log the failure
+    try {
+        IPython.notebook.kernel.execute(code, null, {store_history: false});        
+    }
+    catch (err) {
+        // wait half a second and try one more time.
+        console.log(err);
+        setTimeout( function() { IPython.notebook.kernel.execute(code, null, {store_history: false}); }, 500 );
+    }    
+    return true;
 }
 /**
  * Syntactic sugar for logging error vs. fatal error.
@@ -272,7 +279,6 @@ narrative.init = function() {
                                                     .append($firstShutdownBtn))
                                             .append($reallyShutdownPanel))));
 
-//    var $versionBtn = $('<a href="#">About</a>')
     $('#kb-about-btn').click(function(event) {
                           event.preventDefault();
                           event.stopPropagation();
@@ -282,7 +288,6 @@ narrative.init = function() {
                           }
                       });
 
-//    $('#kb-version-stamp').empty().append($versionBtn);
     $('#notebook').append($versionModal);
     $('[data-toggle="tooltip"]').tooltip()
     /*
@@ -291,12 +296,22 @@ narrative.init = function() {
     var $sidePanel = $('#kb-side-panel').kbaseNarrativeSidePanel({ autorender: false });
 
     var curCell = null;
-    $([IPython.events]).on('select.Cell', function(event, data) {
-        if (curCell && data.cell != this.curCell)
+    var showIPythonCellToolbar = function(cell) {
+        // hide the previously selected cell's toolbar
+        if (curCell && cell != curCell)
             curCell.celltoolbar.hide();
-        curCell = data.cell;
+        curCell = cell;
+        // show the new one
         if (!curCell.metadata['kb-cell'])
-            curCell.celltoolbar.show();
+            curCell.celltoolbar.show();        
+    };
+
+    $([IPython.events]).on('select.Cell', function(event, data) {
+        showIPythonCellToolbar(data.cell);
+    });
+
+    $([IPython.events]).on('create.Cell', function(event, data) {
+        showIPythonCellToolbar(data.cell);
     });
 
     /*