+++ src/notebook/ipython_profiles/profile_narrative/kbase_templates/static/kbase/js/widgets/narrative_core/kbaseNarrativeDataList.js
@@ -12,7 +12,7 @@
 
             ws_url:"https://kbase.us/services/ws",
             landing_page_url: "/functional-site/#/", // !! always include trailing slash
-            default_landing_page_url: "/functional-site/#/ws/json/", // ws_name/obj_name,
+            default_landing_page_url: "/functional-site/#/json/", // ws_name/obj_name,
 
             user_name_fetch_url:"https://kbase.us/services/genome_comparison/users?usernames=",
 
@@ -402,6 +402,10 @@
                                             if (self.ws_name && self.ws) {
                                                 self.ws.get_object_history({ref:object_info[6]+"/"+object_info[0]},
                                                     function(history) {
+                                                        $alertContainer.append($('<div>')
+                                                            .append($('<button>').addClass('kb-data-list-cancel-btn')
+                                                                        .append('Hide History')
+                                                                        .click(function() {$alertContainer.empty();} )));
                                                         history.reverse();
                                                         var $tbl = $('<table>').css({'width':'100%'});
                                                         for(var k=0; k<history.length;k++) {
@@ -410,17 +414,18 @@
                                                                 $revertBtn.tooltip({title:'Current Version', 'container':'body',placement:'bottom'});
                                                             } else {
                                                                 var revertRef = {wsid:history[k][6], objid:history[k][0], ver:history[k][4]};
-                                                                $revertBtn.tooltip({title:'Revert to this version?', 'container':'body',placement:'bottom'})
-                                                                    .click(function() {
-                                                                        self.ws.revert_object(revertRef,
-                                                                            function(reverted_obj_info) {
-                                                                                self.refresh();
-                                                                            }, function(error) {
-                                                                                console.error(error);
-                                                                                $alertContainer.empty();
-                                                                                $alertContainer.append($('<span>').css({'color':'#F44336'}).append("Error! "+error.error.message));
-                                                                            });
-                                                                    })
+                                                                (function(revertRefLocal) {
+                                                                    $revertBtn.tooltip({title:'Revert to this version?', 'container':'body',placement:'bottom'})
+                                                                        .click(function() {
+                                                                            self.ws.revert_object(revertRefLocal,
+                                                                                function(reverted_obj_info) {
+                                                                                    self.refresh();
+                                                                                }, function(error) {
+                                                                                    console.error(error);
+                                                                                    $alertContainer.empty();
+                                                                                    $alertContainer.append($('<span>').css({'color':'#F44336'}).append("Error! "+error.error.message));
+                                                                                });
+                                                                        }); })(revertRef);
                                                             }
                                                             $tbl.append($('<tr>')
                                                                         .append($('<td>').append($revertBtn))
@@ -455,7 +460,23 @@
                                         .append($('<span>').addClass('fa fa-download').css(css))
                                         .click(function(e) {
                                             e.stopPropagation(); $alertContainer.empty();
-                                            $alertContainer.append('Coming soon');
+                                            $alertContainer.append($('<div>')
+                                            		.append($('<button>').addClass('kb-data-list-btn')
+                                                            .append('Download as JSON')
+                                                            .click(function() {
+                                                            	var url = window.kbconfig.urls.data_import_export + '/download?ws='+object_info[7]+'&id='+object_info[1]+'&token='+self._attributes.auth.token;
+                                                            	console.log("Download url=" + url);
+                                                            	var hiddenIFrameID = 'hiddenDownloader';
+                                                                var iframe = document.getElementById(hiddenIFrameID);
+                                                            	if (iframe === null) {
+                                                            		iframe = document.createElement('iframe');
+                                                            		iframe.id = hiddenIFrameID;
+                                                            		iframe.style.display = 'none';
+                                                            		document.body.appendChild(iframe);
+                                                            	}
+                                                            	iframe.src = url;
+                                                            }))
+                                            );
                                         });
 
             var $rename = $('<span>')
@@ -799,6 +820,9 @@
                         if (self.n_objs_rendered >= start+self.options.objs_to_render_on_scroll) {
                             break;
                         }
+                        if (self.objectList[i].key == undefined) {
+                            self.objectList[i].key = self.genUUID();
+                        }
                         self.attachRow(i);
                     }
                     //console.log('showing '+ self.n_objs_rendered + ' of ' + self.objectList.length);
@@ -957,6 +981,7 @@
 
             var $openSearch = $('<span>')
                 .addClass('btn btn-xs btn-default')
+                .tooltip({title:'Search data in narrative', 'container':'body', delay: { "show": 400, "hide": 50 }})
                 .append('<span class="fa fa-search"></span>')
                 .on('click',function() {
                     if(!self.$searchDiv.is(':visible')) {
@@ -970,6 +995,7 @@
 
             var $openSort = $('<span>')
                 .addClass('btn btn-xs btn-default')
+                .tooltip({title:'Sort data list', 'container':'body', delay: { "show": 400, "hide": 50 }})
                 .append('<span class="fa fa-sort-amount-asc"></span>')
                 .on('click',function() {
                     if(!self.$sortByDiv.is(':visible')) {
@@ -983,6 +1009,7 @@
 
             var $openFilter = $('<span>')
                 .addClass('btn btn-xs btn-default')
+                .tooltip({title:'Filter data by type', 'container':'body', delay: { "show": 400, "hide": 50 }})
                 .append('<span class="fa fa-filter"></span>')
                 .on('click',function() {
                     if(!self.$filterTypeDiv.is(':visible')) {