+++ src/notebook/ipython_profiles/profile_narrative/kbase_templates/static/kbase/js/widgets/narrative_core/kbaseNarrativeDataPanel.js
@@ -39,7 +39,7 @@
             workspaceURL: "https://kbase.us/services/ws",
             wsBrowserURL: "/functional-site/#/ws/",
             landing_page_url: "/functional-site/#/", // !! always include trailing slash
-            default_landing_page_url: "/functional-site/#/ws/json/", // ws_name/obj_name,
+            default_landing_page_url: "/functional-site/#/json/", // ws_name/obj_name,
             container: null,
             ws_name: null,
         },
@@ -80,7 +80,6 @@
             
             $(document).on(
                 'setWorkspaceName.Narrative', $.proxy(function(e, info) {
-                    console.log('data panel -- setting ws to ' + info.wsId);
                     this.ws_name = info.wsId;
                     this.narrWs = info.narrController;
                     this.dataListWidget.setWorkspace(this.ws_name);
@@ -137,6 +136,7 @@
 
             this.addButton($('<button>')
                            .addClass('btn btn-xs btn-default')
+                           .tooltip({title:'Hide / Show data browser', 'container':'body', delay: { "show": 400, "hide": 50 }})
                            .append('<span class="fa fa-arrow-right"></span>')
                            .click($.proxy(function(event) {
                                this.trigger('hideGalleryPanelOverlay.Narrative');
@@ -359,8 +359,10 @@
             
             // hack to keep search on top
             var $mineScrollPanel = $('<div>').css({'overflow-x':'hidden','overflow-y':'auto','height':'430px'});
+            setLoading($mineScrollPanel);
             minePanel.append($mineScrollPanel);
             var $sharedScrollPanel = $('<div>').css({'overflow-x':'hidden','overflow-y':'auto','height':'430px'});
+            setLoading($sharedScrollPanel);
             sharedPanel.append($sharedScrollPanel);
             
             // Setup the panels that are defined by widgets
@@ -393,28 +395,9 @@
             // start with my data, then fetch other data
             // this is because data sets can be large and
             // makes things more fluid
-            minePanel.loading();
-            sharedPanel.loading();
             updateView('mine').done(function() {
                 updateView('shared')
             });
-
-            // events for changing tabs
-            /*$($tabs.header.find('.kb-side-header')).click(function() {
-                // reset selected models when changing tabs, if that's what is wanted
-                if ($(this).index() == 0)
-                    mineSelected = [], btn.show();
-                else if ($(this).index() == 1)
-                    sharedSelected = [], btn.show();
-                else
-                    btn.hide();
-
-                // reset checkboxs... for any tabs.
-                var checkboxes = body.find('.kb-import-checkbox');
-                checkboxes.removeClass('fa-check-square-o')
-                          .addClass('fa-square-o');
-                btn.prop('disabled', true);
-            })*/
 
             var narrativeNameLookup={};
             this.$overlayPanel = $('<div>').append(body).append(footer);
@@ -429,10 +412,8 @@
                     else if (view == 'shared') prom = getSharedData(workspaces);
                     $.when(prom).done(function() {
                         if (view == 'mine') {
-                            minePanel.rmLoading();
                             addMyFilters();
                         } else if(view == 'shared') {
-                            sharedPanel.rmLoading();
                             addSharedFilters();
                         }
                     });
@@ -441,6 +422,10 @@
 
             // function used to update my data list
             function getMyData(workspaces, type, ws_name) {
+                if (workspaces.length==0) {
+                    render([], $mineScrollPanel, {});
+                    return [];
+                }
                 var params = {includeMetadata:1};
                 if (!ws_name) {
                     var ws_ids = [], obj_count = 0;
@@ -485,6 +470,10 @@
 
             // function used to update shared with me data list
             function getSharedData(workspaces, type, ws_name) {
+                if (workspaces.length==0) {
+                    render([], $sharedScrollPanel, {});
+                    return null;
+                }
                 var params = {includeMetadata:1};
                 if (!ws_name) {
                     var ws_ids = [], obj_count = 0;
@@ -534,11 +523,11 @@
                 var start = 0, end = 30;
 
                 // remove items from only current container being rendered
-                container.find('.kb-import-items').remove();
+                container.empty();
 
                 if (data.length == 0){
-                    container.append('<div class="kb-import-items text-muted">No data found</div>');
-                    return
+                    container.append($('<div>').addClass("kb-data-list-type").css({margin:'15px', 'margin-left':'35px'}).append('No data found'));
+                    return;
                 } else if (data.length-1 < end)
                     end = data.length;
 
@@ -546,6 +535,11 @@
                 container.append(rows);
                 events(container, selected);
 
+                if (rows.children().length==0) {
+                    container.append($('<div>').addClass("kb-data-list-type").css({margin:'15px', 'margin-left':'35px'}).append('No data found'));
+                    return;
+                }
+                
                 // infinite scroll
                 var currentPos = end;
                 container.unbind('scroll');
@@ -819,6 +813,10 @@
 
                 // possible filters via input
                 var type, ws, query;
+                
+                // create filter (search)
+                var filterInput = $('<input type="text" class="form-control kb-import-search" placeholder="Search data...">');
+                var searchFilter = $('<div class="col-sm-4">').append(filterInput);
 
                 // create workspace filter
                 var wsInput = $('<select class="form-control kb-import-filter">');
@@ -833,13 +831,10 @@
                 // event for type dropdown
                 wsInput.change(function() {
                     ws = $(this).children('option:selected').data('name');
-
+                    filterInput.val('');
                     // request again with filted type
-                    minePanel.find('.kb-import-items').remove();
-                    minePanel.loading();
-                    getMyData(myWorkspaces, type, ws).done(function() {
-                        minePanel.rmLoading();
-                    })
+                    setLoading($mineScrollPanel);
+                    getMyData(myWorkspaces, type, ws);
                 })
 
                 // create type filter
@@ -855,23 +850,18 @@
                 // event for type dropdown
                 typeInput.change(function() {
                     type = $(this).children('option:selected').data('type');
-
+                    filterInput.val('');
                     // request again with filted type
-                    minePanel.find('.kb-import-items').remove();
-                    minePanel.loading();
-                    getMyData(myWorkspaces, type, ws).done(function() {
-                        minePanel.rmLoading();
-                    })
+                    //minePanel.loading(); // loading puts the loading image in the wrong place..
+                    setLoading($mineScrollPanel);
+                    getMyData(myWorkspaces, type, ws);
                 })
 
-
-                // create filter (search)
-                var filterInput = $('<input type="text" class="form-control kb-import-search" placeholder="Search data...">');
-                var searchFilter = $('<div class="col-sm-4">').append(filterInput);
 
                 // event for filter (search)
                 filterInput.keyup(function(e){
                     query = $(this).val();
+                    setLoading($mineScrollPanel);
                     var filtered = filterData(myData, {type: type, ws:ws, query:query})
                     render(filtered, $mineScrollPanel, mineSelected);
                 });
@@ -887,7 +877,11 @@
 
                 // possible filters via input
                 var type, ws, query;
-
+                
+                // create filter (search)
+                var filterInput = $('<input type="text" class="form-control kb-import-search" placeholder="Search data...">');
+                var searchFilter = $('<div class="col-sm-4">').append(filterInput);
+                
                 // create workspace filter
                 var wsInput = $('<select class="form-control kb-import-filter">');
                 wsInput.append('<option>All narratives...</option>');
@@ -900,14 +894,11 @@
 
                 // event for type dropdown
                 wsInput.change(function() {
+                    filterInput.val('');
                     ws = $(this).children('option:selected').data('name');
-
                     // request again with filted type
-                    sharedPanel.find('.kb-import-items').remove();
-                    sharedPanel.loading();
-                    getSharedData(sharedWorkspaces, type, ws).done(function() {
-                        sharedPanel.rmLoading();
-                    })
+                    setLoading($sharedScrollPanel);
+                    getSharedData(sharedWorkspaces, type, ws);
                 })
 
 
@@ -924,24 +915,18 @@
                 // event for type dropdown
                 typeInput.change(function() {
                     type = $(this).children('option:selected').data('type');
+                    filterInput.val('');
 
                     // request again with filted type
-                    sharedPanel.find('.kb-import-items').remove();
-                    sharedPanel.loading();
-                    getSharedData(sharedWorkspaces, type, ws).done(function() {
-                        sharedPanel.rmLoading();
-                    })
+                    setLoading($sharedScrollPanel);
+                    getSharedData(sharedWorkspaces, type, ws);
                 })
 
 
-                // create filter (search)
-                var filterInput = $('<input type="text" class="form-control kb-import-search" placeholder="Search data...">');
-                var searchFilter = $('<div class="col-sm-4">').append(filterInput);
 
                 // event for filter (search)
                 filterInput.keyup(function(e){
                     query = $(this).val();
-
                     var filtered = filterData(sharedData, {type: type, ws:ws, query:query})
                     render(filtered, $sharedScrollPanel, sharedSelected);
                 });
@@ -996,7 +981,16 @@
                 var $version = $('<span>').addClass("kb-data-list-version").append('v'+object_info[4]);
                 var $type = $('<span>').addClass("kb-data-list-type").append(type);
 
-                var $date = $('<span>').addClass("kb-data-list-date").append(getTimeStampStr(object_info[3]) + ' by '+object_info[5]);
+                var $date = $('<span>').addClass("kb-data-list-date").append(getTimeStampStr(object_info[3]));
+                var $byUser = $('<span>').addClass("kb-data-list-edit-by");
+                if (object_info[5] !== self.my_user_id) {
+                    $byUser.append(' by '+object_info[5])
+                        .click(function(e) {
+                            e.stopPropagation();
+                            window.open(self.options.landing_page_url+'people/'+object_info[5]);
+                        });
+                }
+                
                 var metadata = object_info[10];
                 var metadataText = '';
                 for(var key in metadata) {
@@ -1010,31 +1004,36 @@
                     }
                 }
 
-                var $savedByUserSpan = $('<td>').addClass('kb-data-list-username-td').append(object_info[5]);
+                var narName = obj.ws;
+                if (narrativeNameLookup[obj.ws]) {
+                    narName = narrativeNameLookup[obj.ws];
+                }
+                var $narName = $('<span>').addClass("kb-data-list-narrative").append(narName);
+                //var $savedByUserSpan = $('<td>').addClass('kb-data-list-username-td').append(object_info[5]);
                 //this.displayRealName(object_info[5],$savedByUserSpan);
 
-                var typeLink = 'type';//'<a href="'+this.options.landing_page_url+'spec/module/'+type_module+'" target="_blank">' +type_module+"</a>.<wbr>" +
+                //var typeLink = '<a href="'+this.options.landing_page_url+'spec/module/'+type_module+'" target="_blank">' +type_module+"</a>.<wbr>" +
                 //                '<a href="'+this.options.landing_page_url+'spec/type/'+object_info[2]+'" target="_blank">' +(type_tokens[1].replace('-','&#8209;')) + '.' + type_tokens[2] + '</a>';
-                var $moreRow  = $('<div>').addClass("kb-data-list-more-div").hide()
-                                .append(
-                                    $('<table style="width:100%;">')
-                                        .append("<tr><th>Permament Id</th><td>" +object_info[6]+ "/" +object_info[0]+ "/" +object_info[4] + '</td></tr>')
-                                        .append("<tr><th>Full Type</th><td>"+typeLink+'</td></tr>')
-                                        .append($('<tr>').append('<th>Saved by</th>').append($savedByUserSpan))
-                                        .append(metadataText));
+                //var $moreRow  = $('<div>').addClass("kb-data-list-more-div").hide()
+                //                .append(
+                //                    $('<table style="width:100%;">')
+                //                        .append("<tr><th>Permament Id</th><td>" +object_info[6]+ "/" +object_info[0]+ "/" +object_info[4] + '</td></tr>')
+                //                        .append("<tr><th>Full Type</th><td>"+typeLink+'</td></tr>')
+                //                        .append($('<tr>').append('<th>Saved by</th>').append($savedByUserSpan))
+                //                        .append(metadataText));
 
-                var $toggleAdvancedViewBtn = $('<span>').addClass("kb-data-list-more")//.addClass('btn btn-default btn-xs kb-data-list-more-btn')
-                    .hide()
-                    .html('<span class="fa fa-ellipsis-h" style="color:#999" aria-hidden="true"/>');
-                var toggleAdvanced = function() {
-                        if ($moreRow.is(':visible')) {
-                            $moreRow.slideUp('fast');
-                            $toggleAdvancedViewBtn.show();
-                        } else {
-                            $moreRow.slideDown('fast');
-                            $toggleAdvancedViewBtn.hide();
-                        }
-                    };
+                //var $toggleAdvancedViewBtn = $('<span>').addClass("kb-data-list-more")//.addClass('btn btn-default btn-xs kb-data-list-more-btn')
+                //    .hide()
+                //    .html('<span class="fa fa-ellipsis-h" style="color:#999" aria-hidden="true"/>');
+                //var toggleAdvanced = function() {
+                //        if ($moreRow.is(':visible')) {
+                //            $moreRow.slideUp('fast');
+                //            $toggleAdvancedViewBtn.show();
+                //        } else {
+                //            $moreRow.slideDown('fast');
+                //            $toggleAdvancedViewBtn.hide();
+                //        }
+                //    };
     
                 var $btnToolbar = $('<span>').addClass('btn-toolbar pull-right').attr('role', 'toolbar').hide();
                 var btnClasses = "btn btn-xs btn-default";
@@ -1064,8 +1063,8 @@
                 var $mainDiv  = $('<div>').addClass('kb-data-list-info').css({padding:'0px',margin:'0px'})
                                     .append($btnToolbar)
                                     .append($name).append($version).append('<br>')
-                                    .append($type).append('<br>').append($date)
-                                    .append($toggleAdvancedViewBtn)
+                                    .append($type).append('<br>').append($narName).append('<br>').append($date).append($byUser);
+                                    //.append($toggleAdvancedViewBtn)
                                     //.click(
                                     //    function() {
                                     //        toggleAdvanced();
@@ -1115,7 +1114,7 @@
                                 //.addClass('kb-data-list-obj-row')
                                 .append($('<div>').addClass('kb-data-list-obj-row-main')
                                             .append($topTable))
-                                .append($moreRow)
+                                //.append($moreRow)
                                 // show/hide ellipses on hover, show extra info on click
                                 .mouseenter(function(){
                                     //if (!$moreRow.is(':visible')) { $toggleAdvancedViewBtn.show(); }
@@ -1135,63 +1134,16 @@
                                                 .css({'margin-left':'150px'}))
                                     .append($row);
     
-    
-                //item.append('<i class="fa fa-square-o pull-left kb-import-checkbox">');
                 return $rowWithHr;
-            
-                
-                
-                
-               
-                
-                
-                /*var item = $('<div class="kb-import-item">')
-                                .data('ref', obj.wsID+'.'+obj.id)
-                                .data('obj-name', obj.name);
-                item.append('<i class="fa fa-square-o pull-left kb-import-checkbox">');
-                var linkUrl = objURL(obj.module, obj.kind, obj.ws, obj.name);
-                //console.log(linkUrl);
-                item.append('<a class="h4" href="'+
-                                linkUrl+
-                                '" target="_blank">'+obj.name+'</a>'+
-                            '<span class="kb-data-list-version">v'+obj.version+'</span>');
-
-                item.append('<br>');
-
-                item.append('<div class="kb-import-info">'+
-                                '<span>TYPE</span><br>'+
-                                '<b>'+obj.kind+'</b>'+
-                            '</div>');
-                var narName = obj.ws;
-                if (narrativeNameLookup[obj.ws]) {
-                    narName = narrativeNameLookup[obj.ws];
-                }
-                item.append('<div class="kb-import-info">'+
-                                '<span>NARRATIVE</span><br>'+
-                                '<b>'+narName+'<b>'+   //<a class="" href="'+wsURL(obj.ws)+'">'
-                            '</div>');
-                item.append('<div class="kb-import-info">'+
-                                '<span>MODIFIED BY</span><br>'+
-                                '<b>'+obj.info[5]+'<b>'+   //<a class="" href="'+wsURL(obj.ws)+'">'
-                            '</div>');
-                item.append('<div class="kb-import-info">'+
-                                '<span>LAST MODIFIED</span><br>'+
-                                '<b>'+obj.relativeTime+'</b>'+
-                            '</div>');
-                item.append('<br><hr>')
-
-                return item;*/
             }
             
+            // the existing .loading() .rmLoading() puts the loading icon in the wrong place
+            function setLoading($container) {
+                $container.empty();
+                $container.append($('<div>').addClass("kb-data-list-type").css({margin:'15px', 'margin-left':'35px'})
+                                  .append('<img src="' + self.options.loadingImage + '">'));
+            }
             
-            
-            
-            
-            
-            
-            
-            
-
             function objURL(module, type, ws, name) {
                 var mapping = window.kbconfig.landing_page_map;
                 if (mapping[module] && mapping[module][type]) {